Created GWT-specific patching commands.
At the moment there are no commands implemented
so those files just server as a template for
future command implementations.
